Time: 102 ms (6.61%), Space: 115.1 MB (8.28%) - LeetHub

+++ 1838-frequency-of-the-most-frequent-element/1838-frequency-of-the-most-frequent-element.cpp
@@ -12,7 +12,7 @@ public:
             int mid = ( right + left ) / 2;
 
             long actsum = presum[target_idx] - presum[mid] + nums[mid];
-            long sum = (target_idx - mid + 1) * target;
+            long sum = (long)(target_idx - mid + 1) * target;
 
             int opp = sum - actsum;
